LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-fr-expected.txt

PASS navigator.language is "fr-fr"
PASS successfullyParsed is true

TEST COMPLETE
